+++ CHANGELOG.md
@@ -76,6 +76,7 @@ For older changes see the [archived Singularity change log](https://github.com/a
     source. *Applies to newly built containers only, use `apptainer inspect`
     to check version that container was built with*.
 - Added `--no-eval` to the list of flags set by the OCI/Docker `--compat` mode.
+- `sinit` process has been renamed to `appinit`.
 
 ### New features / functionalities